+++ src/main/java/in/ashokit/repository/CustomerRepository.java
@@ -3,7 +3,7 @@ package in.ashokit.repository;
 import in.ashokit.entity.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerEntity,Long> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity,Integer> {
 
     public CustomerEntity findByEmailAndPwd(String email, String pwd);
